more validation for data.py
